fix: unbreak Linux CI — split Windows-only coverage tests
coverage_gaps_test.go referenced pathsEqualCI / historicalPowerShell*
symbols that only exist under //go:build windows, so ubuntu lint and
tests failed to compile. Move those cases to
coverage_gaps_windows_test.go.

Also stop counting stale-host strip paths as "installed", and word apply
messaging as "Updated" so cleanup-only runs do not look like installs.
